Add basic EventLogging to Wikimedia Blog
This change set adds an image tag to the Wikimedia blog side-bar that logs the
request and referer(sic) URLs using the bits endpoint for EventLogging. As
requested by Diederik and Tilman.

I tested the code-block within the get_eventlogging_url() by itself, but I
don't have a full-fledged WMF Wordpress setup to experiment with to verify the
whole package.

Change-Id: I9bd1d3f53e1cb7ee54a0c7abbd2d2ccf40f946e3